+++ src/basic/main.basic.js
@@ -31,9 +31,17 @@ import {
   formatProductDisplayText,
   applyBestBulkDiscount,
   isValidUpdatedQuantity,
+  updateProductOption,
+  renderBonusPoints,
+  updateStockStatus,
+  calculateCart,
 } from './utils.js';
 
 import { state } from './state.js';
+import {
+  handleAddCartClick,
+  handleCartItemClick,
+} from './handlers/cartHandlers.js';
 
 /**
  * 진행 순서
@@ -50,9 +58,24 @@ import { state } from './state.js';
  * 11. 전역변수 순차적 정리
  */
 
+export const dom = {
+  select: null,
+  addCartButton: null,
+  cartDisplay: null,
+  totalDisplay: null,
+  stockStatus: null,
+};
+
+export const cartDisplay = (dom.cartDisplay = document.createElement('div'));
+export const totalDisplay = (dom.totalDisplay = document.createElement('div'));
+export const select = (dom.select = document.createElement('select'));
+export const addCartButton = (dom.addCartButton =
+  document.createElement('button'));
+export const stockStatus = (dom.stockStatus = document.createElement('div'));
+
 // main 함수 - DOM 구성 및 이벤트 설정
 const main = () => {
-  let select, addCartButton, cartDisplay, totalDisplay, stockStatus; // DOM 요소들
+  // let select, addCartButton, cartDisplay, totalDisplay, stockStatus; // DOM 요소들
 
   // 상품 데이터
   const prodList = (state.prodList = [
@@ -68,12 +91,6 @@ const main = () => {
   const container = document.createElement('div');
   const wrapper = document.createElement('div');
   const title = document.createElement('h1');
-
-  cartDisplay = document.createElement('div');
-  totalDisplay = document.createElement('div');
-  select = document.createElement('select');
-  addCartButton = document.createElement('button');
-  stockStatus = document.createElement('div');
 
   // 클래스 및 ID 설정
   cartDisplay.id = 'cart-items';
@@ -106,135 +123,6 @@ const main = () => {
   root
     ? root.appendChild(container)
     : console.error("Element with ID 'app' not found in the DOM.");
-
-  // event listeners
-
-  /**
-   * 상품 선택 옵션 업데이트
-   */
-  function updateProductOption() {
-    select.innerHTML = '';
-
-    prodList.forEach((item) => {
-      const option = document.createElement('option');
-      option.value = item.id;
-      option.textContent = item.name + ' - ' + item.val + '원';
-
-      if (item.q === 0) option.disabled = true;
-
-      select.appendChild(option);
-    });
-  }
-
-  /**
-   * 장바구니 계산 함수
-   */
-  function calculateCart() {
-    state.finalPaymentAmount = 0; // 최종 결제 금액
-    state.cartAllItemCount = 0; // 카트 전체 상품 갯수
-    const cartItems = cartDisplay.children; // 장바구니에 담긴 상품들
-    let cartAllItemPrice = 0; // 카트 전체 상품 가격
-
-    // 장바구니에 담긴 상품들
-    for (const cartItem of cartItems) {
-      const productId = cartItem.id;
-      const currentProduct = getProductById(prodList, productId);
-      const quantity = getQuantityFromText(
-        cartItem.querySelector('span').textContent,
-        QUANTITY_TEXT_SEPARATOR,
-      );
-
-      const productPrice = currentProduct.val * quantity;
-      const productDiscount =
-        quantity >= BULE_PURCHASE_PRODUCT_THRESHOLD
-          ? PRODUCT_DISCOUT_MAP[currentProduct.id] || 0
-          : 0;
-
-      state.cartAllItemCount += quantity;
-      cartAllItemPrice += productPrice;
-      state.finalPaymentAmount += getDiscountPrice(
-        productPrice,
-        productDiscount,
-      );
-    }
-
-    let discountPercent = 0; // 할인율
-
-    // 전체 수량이 30개 이상일때 적용할 수 있는 대량 구매 할인
-    if (state.cartAllItemCount >= BULK_PURCHASE_THRESHOLD) {
-      const { price, rate } = applyBestBulkDiscount(
-        cartAllItemPrice,
-        state.finalPaymentAmount,
-        BULK_PURCHASE_THRESHOLD,
-        EXTRA_DISCOUT_PERCENT,
-      );
-      state.finalPaymentAmount = price;
-      discountPercent = rate;
-    } else {
-      discountPercent =
-        (cartAllItemPrice - state.finalPaymentAmount) / cartAllItemPrice;
-    }
-
-    // 화요일 할인 적용
-    if (isTuesday()) {
-      state.finalPaymentAmount *= 1 - TUESDAY_DISCOUNT_AMOUNT_PERCENT;
-      discountPercent = Math.max(
-        discountPercent,
-        TUESDAY_DISCOUNT_AMOUNT_PERCENT,
-      );
-    }
-
-    // 할인율이 0보다 클 경우
-    totalDisplay.textContent =
-      '총액: ' + Math.round(state.finalPaymentAmount) + '원';
-
-    // 할인율이 0보다 클 경우
-    if (discountPercent > 0) {
-      const span = document.createElement('span');
-      span.className = CLASS_DISCOUNT_TAG;
-      span.textContent =
-        '(' + (discountPercent * 100).toFixed(1) + '% 할인 적용)';
-      totalDisplay.appendChild(span);
-    }
-
-    // 재고 정보 업데이트
-    updateStockStatus();
-    renderBonusPoints();
-  }
-
-  /**
-   * 포인트 랜더링
-   */
-  const renderBonusPoints = () => {
-    state.bonusPoints = Math.floor(state.finalPaymentAmount / 1000);
-    let pointsTag = document.getElementById('loyalty-points');
-    if (!pointsTag) {
-      pointsTag = document.createElement('span');
-      pointsTag.id = 'loyalty-points';
-      pointsTag.className = CLASS_POINTS_TAG;
-      totalDisplay.appendChild(pointsTag);
-    }
-    pointsTag.textContent = '(포인트: ' + state.bonusPoints + ')';
-  };
-
-  /**
-   * 재고 정보 업데이트
-   */
-  const updateStockStatus = () => {
-    let infoMsg = '';
-    prodList.forEach((item) => {
-      if (item.q < 5) {
-        infoMsg +=
-          item.name +
-          ': ' +
-          (item.q > 0 ? '재고 부족 (' + item.q + '개 남음)' : '품절') +
-          '\n';
-      }
-    });
-    stockStatus.textContent = infoMsg;
-  };
-
-  // set up event listeners
 
   /**
    * 번개세일 타이머 설정
@@ -275,135 +163,8 @@ const main = () => {
   }, Math.random() * 20000);
 
   // Initial rendering
-  addCartButton.addEventListener('click', () => {
-    const selectedProductId = select.value; // 선택된 상품 ID
-    const itemToAdd = getProductById(prodList, selectedProductId); // 선택된 상품
-
-    if (itemToAdd && itemToAdd.q > 0) {
-      const item = document.getElementById(itemToAdd.id);
-
-      if (item && item !== null) {
-        // 이미 장바구니에 있는 상품일 경우
-        const updatedQuantity = getUpdatedQuantity(
-          item.querySelector('span').textContent,
-          1,
-          QUANTITY_TEXT_SEPARATOR,
-        );
-
-        if (updatedQuantity <= itemToAdd.q) {
-          const spanElement = item.querySelector('span');
-
-          // 수량 업데이트
-          if (spanElement) {
-            spanElement.textContent = formatProductDisplayText(
-              itemToAdd.name,
-              itemToAdd.val,
-              updatedQuantity,
-              QUANTITY_TEXT_SEPARATOR,
-            );
-          } else {
-            console.error('Span element not found for item:', itemToAdd.id);
-          }
-          itemToAdd.q--;
-        } else {
-          alert('재고가 부족합니다.');
-        }
-      } else {
-        const newItem = document.createElement('div');
-        newItem.id = itemToAdd.id;
-        newItem.className = CLASS_PRODUCT_WRAPPER;
-        newItem.innerHTML =
-          `<span>${formatProductDisplayText(itemToAdd.name, itemToAdd.val, 1, QUANTITY_TEXT_SEPARATOR)}</span><div>` +
-          `<button class="${CLASS_BTN_QUANTITY}" data-product-id="` +
-          itemToAdd.id +
-          '" data-change="-1">-</button>' +
-          `<button class="${CLASS_BTN_QUANTITY}" data-product-id="` +
-          itemToAdd.id +
-          '" data-change="1">+</button>' +
-          `<button class="${CLASS_BTN_REMOVE}" data-product-id="` +
-          itemToAdd.id +
-          '">삭제</button></div>';
-        cartDisplay.appendChild(newItem);
-        itemToAdd.q--;
-      }
-      calculateCart();
-      state.lastSelectedProductId = selectedProductId;
-    }
-  });
-
-  /**
-   * 장바구니 아이템 수량 변경 및 삭제 이벤트 핸들러 (이벤트 위임)
-   */
-  cartDisplay.addEventListener('click', (event) => {
-    const target = event.target;
-
-    if (
-      (target instanceof HTMLElement &&
-        target.classList.contains('quantity-change')) ||
-      (target instanceof HTMLElement &&
-        target.classList.contains('remove-item'))
-    ) {
-      const targetProductId = target.dataset[DATASET_KEY_PRODUCT_ID]; // target
-      const itemElement = targetProductId
-        ? document.getElementById(targetProductId)
-        : null;
-      const selectedProduct = getProductById(prodList, targetProductId); // 상품 찾기
-
-      // 이벤트 델리게이션이나 특정 요소만 조건 분기할 때 자주 씀
-      if (
-        target.classList.contains('quantity-change') &&
-        selectedProduct &&
-        itemElement &&
-        itemElement.querySelector('span')?.textContent
-      ) {
-        // 수량 변경 버튼 클릭
-        const currentQuantity = parseInt(
-          target.dataset[DATASET_KEY_CHANGE] || '0',
-        );
-        // 수량 변경
-        const updatedQuantity = getUpdatedQuantity(
-          itemElement.querySelector('span').textContent,
-          currentQuantity,
-          QUANTITY_TEXT_SEPARATOR,
-        );
-        if (
-          isValidUpdatedQuantity(
-            updatedQuantity,
-            selectedProduct.q,
-            getQuantityFromText(
-              itemElement.querySelector('span').textContent,
-              QUANTITY_TEXT_SEPARATOR,
-            ),
-          )
-        ) {
-          itemElement.querySelector('span').textContent =
-            formatProductDisplayText(
-              selectedProduct.name,
-              selectedProduct.val,
-              updatedQuantity,
-              QUANTITY_TEXT_SEPARATOR,
-            );
-
-          selectedProduct.q -= currentQuantity;
-        } else if (updatedQuantity <= 0) {
-          itemElement.remove();
-          selectedProduct.q -= currentQuantity;
-        } else {
-          alert('재고가 부족합니다.');
-        }
-      } else if (target.classList.contains('remove-item')) {
-        const selectedQuantity = parseInt(
-          // 남은 수량
-          itemElement
-            .querySelector('span')
-            .textContent.split(QUANTITY_TEXT_SEPARATOR)[1],
-        );
-        selectedProduct.q += selectedQuantity;
-        itemElement.remove();
-      }
-      calculateCart();
-    }
-  });
+  addCartButton.addEventListener('click', () => handleAddCartClick());
+  cartDisplay.addEventListener('click', (event) => handleCartItemClick(event));
 
   updateProductOption(); // ✅ 초기 옵션 렌더링
   calculateCart(); // ✅ 초기 총액 렌더링
